feat: core 파일 *.d.ts 파일 추가
