+++ src/pages/type/index.ts
@@ -1,4 +1,4 @@
-import Component from '@core/Component';
+import Component from '@core/component';
 import './style.scss';
 
 export default class Type extends Component {
