+++ src/style.scss
@@ -1,2 +1,3 @@
 @import './stylesheets/color.scss';
 @import './stylesheets/mixin.scss';
+@import './stylesheets/reset.scss';
