+++ src/types/observer.d.ts
@@ -1,0 +1,11 @@
+interface Fn {
+  (): void;
+}
+
+interface ObserverMap<T> {
+  [key: string]: T;
+}
+
+interface Object {
+  [key: string]: any;
+}
