+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "lib": ["es5", "es6", "es2017", "dom"],
     "module": "ES6", // 모듈 형식 ES6
     "moduleResolution": "Node", // 모듈 해석 Node
     "esModuleInterop": true, // commonJS 모듈을 ES6로 import
@@ -7,11 +8,13 @@
     "noImplicitAny": true,
     "strictNullChecks": true,
     "baseUrl": "./",
+    "typeRoots": ["./src/types"],
     "paths": {
       "@core/*": ["src/core/*"],
       "@components/*": ["src/components/*"],
       "@pages/*": ["src/pages/*"],
-      "@stylesheets/*": ["src/stylesheets/*"]
+      "@stylesheets/*": ["src/stylesheets/*"],
+      "@util/*": ["src/util/*"]
     }
   },
   "include": ["src/**/*.ts"]
